test(코멘트 업데이트 테스트 코드 작성): 코멘트 업데이트 테스트 코드 작성
댓글 수정 API 작성.
