--- src/test/java/com/stock/togetherStock/comment/CommentServiceTest.java
@@ -1,52 +1,0 @@
-package com.stock.togetherStock.comment;
-
-import static org.mockito.BDDMockito.given;
-
-import com.stock.togetherStock.comment.repository.CommentRepository;
-import com.stock.togetherStock.comment.service.CommentService;
-import com.stock.togetherStock.member.domain.MemberDto;
-import com.stock.togetherStock.post.domain.Post;
-import com.stock.togetherStock.post.domain.PostDto;
-import com.stock.togetherStock.post.service.PostService;
-import java.time.LocalDateTime;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-public class CommentServiceTest {
-
-    @Mock
-    CommentRepository commentRepository;
-
-    @InjectMocks
-    CommentService commentService;
-
-    PostService postService;
-
-    protected MemberDto createMemberDto() {
-
-        return MemberDto.builder()
-            .email("test@naver.com")
-            .password("12345")
-            .name("test")
-            .nickname("hateTest")
-            .phone("11111111")
-            .build();
-    }
-
-    protected PostDto createPostDto() {
-        return PostDto.builder()
-            .title("제목1")
-            .content("내용1")
-            .regiPostDate(LocalDateTime.now())
-            .build();
-    }
-
-
-
-
-}
